добавляет выполнение задания 5.14
